Clear history shortcut added - Created ClearHistoryHelper to do that
